# aeterna-openapi-spec.yaml
openapi: '3.1.0'
info:
  title: Aeterna Admin API
  version: 0.8.0
  description: |
    REST API for the Aeterna memory and knowledge platform.
    All list endpoints return a `PaginatedResponse` envelope.
servers:
  - url: http://localhost:7777/api/v1
    description: Local dev
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...
  schemas:
...
    MemoryEntry:
      type: object
      properties:
        id: { type: string }
        content: { type: string }
        layer: { type: string }
        importanceScore: { type: number }
        metadata: { type: object }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
security:
  - bearerAuth: []
paths:
...
  /memory/search:
    post:
      summary: Semantic memory search
      tags: [Memory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [query]
              properties:
                query: { type: string }
                limit: { type: integer, default: 10 }
                offset: { type: integer, default: 0 }
                threshold: { type: number, default: 0 }
                filters: { type: object }
                contextSummary: { type: string }
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/MemoryEntry' } }
                  total: { type: integer }
                  reasoning: { type: string, nullable: true }
  /memory/list:
    post:
      summary: List memory entries from a layer
      tags: [Memory]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [layer]
              properties:
                layer: { type: string }
                limit: { type: integer, default: 50 }
                offset: { type: integer, default: 0 }
      responses:
        '200':
          description: Memory entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/MemoryEntry' } }
                  total: { type: integer }
...